Small update

+++ resources/views/articles/partials/article-list.blade.php
@@ -1,22 +1,30 @@
 <article data-article="{{ $post->id }}" class="article col-xs-12">
     <h2>{{ $post->title }}</h2>
-    <hr>
-    <div class="article-property col-xs-12 col-sm-4 col-md-2">
-        <div class="article-date">{{ $post->created_at->diffForHumans() }}</div>
-        <div class="article-tags">
-            @each('articles.partials.tag', $post->tags, 'tag')
-        </div>
+    <div class="article-properties">
+        <span class="article-property" data-toggle="tooltip"
+              title="{{ $post->created_at->format("d. m. Y H:i") }}" data-placement="bottom">{{ $post->created_at->diffForHumans() }}</span>
+        <span class="article-property">
+            @foreach($post->tags as $tag)
+                <a href="{{ url('tags/'.$tag->name) }}">{{ $tag->name }}</a>@unless($loop->last), @endunless
+            @endforeach
+        </span>
+        <span class="article-property">by Žiga Strgar</span>
     </div>
-    <div class="col-xs-12 col-sm-8 col-md-10">
+    <hr>
+    <div class="col-xs-12">
         <div class="article-content">
-            {{ $post->content }}
+            {{ $post->excerpt }}
         </div>
-        <a class="pull-right btn btn--border-danger btn--bg-danger"
-           onClick="deleteArticle({{ $post->id }});">Delete article</a>
-        <a class="pull-right btn btn--border-blue btn--bg-transparent"
-           href="{{ action('ArticlesController@edit', ['id' => $post->id]) }}">Edit article</a>
-        <a class="pull-right btn btn--border-blue btn--bg-transparent"
-           href="{{ action('ArticlesController@show', ['id' => $post->id]) }}">Read more ...</a>
+        <div class="article-action">
+            <a class="btn btn--border-blue btn--bg-transparent"
+               href="{{ action('ArticlesController@show', ['id' => $post->slug]) }}">Read more ...</a>
+            @if(Auth::check() && Auth::user()->isAdmin())
+                <a class="btn btn--border-blue btn--bg-transparent"
+                   href="{{ action('ArticlesController@edit', ['id' => $post->id]) }}">Edit article</a>
+                <a class="btn btn--border-danger btn--bg-danger"
+                   onClick="deleteArticle({{ $post->id }});">Delete article</a>
+            @endif
+        </div>
     </div>
 </article>
 
@@ -33,6 +41,10 @@
                     }
                 })
             }
+
+            $(function () {
+                $("[data-toggle=tooltip]").tooltip();
+            });
         </script>
     @endif
 @endsection
